Adds a sound when ai finds a set
Gives auditory feedback to the user when the ai has found a set so they have extra signal to stop actively looking.  Uses sound 3 which is used for selecting and incorrect set and losing the game as it indicates "bad" which is appropriate here for the ai opponent earning a point.